Refactor: Quantity double 로 변경
주식 수량을 정하는 quantity 필드가 int 로 되어 있던것을 double 로 변경하였습니다.

related to: #51

+++ src/main/java/com/prgrms/ijuju/domain/stock/adv/advancedinvest/dto/request/StockTransactionRequestDto.java
@@ -14,7 +14,7 @@ import java.math.BigDecimal;
 @NoArgsConstructor
 public class StockTransactionRequestDto {
     private String stockSymbol;    // 주식 종목
-    private int quantity;          // 구매 또는 판매 수량
+    private double quantity;          // 구매 또는 판매 수량
     private Long memberId;         // 사용자 ID
     private BigDecimal points;     // 거래 금액
     private TradeType tradeType;   // 거래 유형 (BUY or SELL)
